wine_quality_heroku/requirements.txt
click==8.0.3
Flask==2.0.2
Flask-MySQLdb==0.2.0
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
MarkupSafe==2.0.1
mysqlclient==2.1.0
numpy==1.21.4
pandas==1.3.4
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
scikit-learn==1.0.1
scipy==1.7.3
six==1.16.0
sklearn==0.0
threadpoolctl==3.0.0
Werkzeug==2.0.2
